CPaperUI/requirements.txt
pandas
streamlit-option-menu
python-docx